[ADS-278] Expiration of Request
Main changes:
- Check expiry against requestTime and signatureTime
- Set default expiration time to 60s

Other changes:
- Refactor SignatureVerifier, encapsulate code and leave only public API methods in SignatureVerifier
- Move examples from this repo to separate one and add submodule
