+++ AdScore.Signature.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.30804.86
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdScore.Signature", "src\AdScore.Signature.csproj", "{B3C2D7CD-CC89-4CA4-9395-723F489D3052}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleApp", "sample\ExampleApp\ExampleApp.csproj", "{DB14479E-B595-4635-9425-8C3B0E93FD42}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{B3C2D7CD-CC89-4CA4-9395-723F489D3052}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B3C2D7CD-CC89-4CA4-9395-723F489D3052}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B3C2D7CD-CC89-4CA4-9395-723F489D3052}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DB14479E-B595-4635-9425-8C3B0E93FD42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DB14479E-B595-4635-9425-8C3B0E93FD42}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DB14479E-B595-4635-9425-8C3B0E93FD42}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DB14479E-B595-4635-9425-8C3B0E93FD42}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
