--- sample/ExampleApp/ExampleApp.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AdScore.Signature" Version="1.0.2" />
-  </ItemGroup>
-
-
-</Project>
