+++ src/AdScore.Signature.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <PackageId>AdScore.Signature</PackageId>
-    <Version>1.0.2</Version>
+    <Version>1.1.0</Version>
     <TargetFramework>netstandard1.6</TargetFramework>
     <RootNamespace>AdScore.Signature</RootNamespace>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
